+++ src/main/java/team03/mopl/domain/watchroom/controller/WatchRoomWebSocketController.java
@@ -8,11 +8,11 @@ import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
-import team03.mopl.domain.watchroom.dto.WatchRoomMessageCreateRequest;
-import team03.mopl.domain.watchroom.dto.WatchRoomMessageDto;
-import team03.mopl.domain.watchroom.dto.ParticipantsInfoDto;
-import team03.mopl.domain.watchroom.dto.VideoControlRequest;
-import team03.mopl.domain.watchroom.dto.VideoSyncDto;
+import team03.mopl.domain.watchroom.dto.watchroommessage.WatchRoomMessageCreateRequest;
+import team03.mopl.domain.watchroom.dto.watchroommessage.WatchRoomMessageDto;
+import team03.mopl.domain.watchroom.dto.participant.ParticipantsInfoDto;
+import team03.mopl.domain.watchroom.dto.video.VideoControlRequest;
+import team03.mopl.domain.watchroom.dto.video.VideoSyncDto;
 import team03.mopl.domain.watchroom.dto.WatchRoomInfoDto;
 import team03.mopl.domain.watchroom.service.WatchRoomMessageService;
 import team03.mopl.domain.watchroom.service.WatchRoomService;
