+++ src/test/java/team03/mopl/domain/watchroom/service/WatchRoomServiceImplTest.java
@@ -25,10 +25,10 @@ import team03.mopl.common.exception.user.UserNotFoundException;
 import team03.mopl.domain.watchroom.dto.WatchRoomContentWithHeadcountDto;
 import team03.mopl.domain.watchroom.dto.WatchRoomCreateRequest;
 import team03.mopl.domain.watchroom.dto.WatchRoomDto;
-import team03.mopl.domain.watchroom.dto.ParticipantDto;
-import team03.mopl.domain.watchroom.dto.ParticipantsInfoDto;
-import team03.mopl.domain.watchroom.dto.VideoControlRequest;
-import team03.mopl.domain.watchroom.dto.VideoSyncDto;
+import team03.mopl.domain.watchroom.dto.participant.ParticipantDto;
+import team03.mopl.domain.watchroom.dto.participant.ParticipantsInfoDto;
+import team03.mopl.domain.watchroom.dto.video.VideoControlRequest;
+import team03.mopl.domain.watchroom.dto.video.VideoSyncDto;
 import team03.mopl.domain.watchroom.dto.WatchRoomInfoDto;
 import team03.mopl.domain.watchroom.entity.WatchRoom;
 import team03.mopl.domain.watchroom.entity.WatchRoomParticipant;
